requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
conda==4.3.16
distlib==0.3.1
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
django-filter==2.4.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
more-itertools==8.2.0
packaging==20.3
pipenv==2020.8.13
pluggy==0.13.1
py==1.8.1
pycosat==0.6.3
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2020.1
psycopg2-binary==2.8.5
requests==2.23.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.25.9
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.1.9
zipp==3.1.0
python-dotenv~=0.15.0
pillow==7.0.0
